drivers: media: Fix -Wpointer-to-int-cast.
drivers/media/platform/msm/cvp/hfi_response_handler.c:507:36: warning: cast to smaller integer type 'unsigned int' from 'void *' [-Wvoid-pointer-to-int-cast]
        inst = cvp_get_inst_from_id(core, (unsigned int)session_id);
                                          ^~~~~~~~~~~~~~~~~~~~~~~~

+++ drivers/media/platform/msm/cvp/hfi_response_handler.c
@@ -504,7 +504,7 @@ static int hfi_process_session_cvp_msg(u32 device_id,
 	}
 	session_id = (void *)(uintptr_t)get_msg_session_id(pkt);
 	core = list_first_entry(&cvp_driver->cores, struct msm_cvp_core, list);
-	inst = cvp_get_inst_from_id(core, (unsigned int)session_id);
+	inst = cvp_get_inst_from_id(core, (uintptr_t)session_id);
 
 	if (!inst) {
 		dprintk(CVP_ERR, "%s: invalid session\n", __func__);
